+++ idea/testData/refactoring/move/kotlin/moveNestedClass/dropEmptyCompanion/before/test.kt
@@ -1,0 +1,7 @@
+package test
+
+class Test1 {
+    companion object {
+        class <caret>Nested{}
+    }
+}